+++ cpack/pack.c
@@ -71,7 +71,7 @@ bool pack_read_bytes(packer_t *pk, uint8_t *result, size_t count)/*{{{*/
 }/*}}}*/
 bool pack_read_string(packer_t *pk, uint8_t **result, size_t *length, alloc_t *alloc)/*{{{*/
 {
-  size_t m;
+  uint64_t m;
   
   *result = 0;
